refactoring de codigo duplicado entre RenderCard y RenderCartCard
